+++ lib/src/gutter.dart
@@ -78,7 +78,8 @@ class Gutter extends StatelessWidget {
       key: key,
       builder: (BuildContext context, Orientation orientation) {
         final double gapSize =
-            size ?? context.gutter(type: type, scaleFactor: scaleFactor);
+            calculateGapSize(context, size, type, scaleFactor);
+
         if (type == GutterType.expand) {
           return Expanded(
             child: SizedBox(
@@ -95,6 +96,14 @@ class Gutter extends StatelessWidget {
       },
     );
   }
+}
+
+/// calculates the gap size based on the size, type and scale factor.
+double calculateGapSize(
+    BuildContext context, double? size, GutterType type, double scaleFactor) {
+  return size != null
+      ? (size * scaleFactor)
+      : context.gutter(type: type, scaleFactor: scaleFactor);
 }
 
 /// A gap a quarter the standard gutter size according to Material Design's
@@ -129,7 +138,89 @@ class GutterExtraLarge extends Gutter {
 /// system
 class Margin extends Gutter {
   /// Creates a new [Margin] widget.
-  const Margin({super.key});
+  const Margin({super.key, super.size, super.type, super.scaleFactor});
+}
+
+/// A widget that insets its child by the material padding or the given padding.
+class GutterPadding extends Gutter {
+  /// Creates a new [GutterPadding] widget with custom padding.
+  const GutterPadding({
+    super.key,
+    this.child,
+    super.size,
+    super.type = GutterType.medium,
+    super.scaleFactor = 1,
+  })  : left = null,
+        top = null,
+        right = null,
+        bottom = null;
+
+  /// Creates a new [GutterPadding] widget with specific edge insets.
+  const GutterPadding.only({
+    super.key,
+    this.child,
+    this.left = 0,
+    this.top = 0,
+    this.right = 0,
+    this.bottom = 0,
+    super.type = GutterType.medium,
+    super.scaleFactor = 1,
+  });
+
+  /// Creates a new [GutterPadding] widget with symmetric padding.
+  const GutterPadding.symmetric({
+    super.key,
+    this.child,
+    double? vertical = 0,
+    double? horizontal = 0,
+    super.type = GutterType.medium,
+    super.scaleFactor = 1,
+  })  : left = horizontal,
+        top = vertical,
+        right = horizontal,
+        bottom = vertical;
+
+  /// Creates a new [GutterPadding] widget with uniform padding.
+  const GutterPadding.all(
+    double? padding, {
+    super.key,
+    this.child,
+    super.type = GutterType.medium,
+    super.scaleFactor = 1,
+  })  : left = padding,
+        top = padding,
+        right = padding,
+        bottom = padding;
+
+  /// The child widget.
+  final Widget? child;
+
+  /// The offset from the left.
+  final double? left;
+
+  /// The offset from the right.
+  final double? right;
+
+  /// The offset from the top.
+  final double? top;
+
+  /// The offset from the bottom.
+  final double? bottom;
+
+  @override
+  Widget build(BuildContext context) {
+    final double gapSize = calculateGapSize(context, size, type, scaleFactor);
+
+    return Padding(
+      padding: EdgeInsets.only(
+        left: left ?? gapSize,
+        top: top ?? gapSize,
+        right: right ?? gapSize,
+        bottom: bottom ?? gapSize,
+      ),
+      child: child,
+    );
+  }
 }
 
 /// An inherited widget that sets custom [Gutter] and [Gap] behavior.